+++ Cargo.toml
@@ -1,9 +1,9 @@
 [workspace]
 members = [
-    "crates/ml", 
-    "crates/engine", 
     "crates/sim_core",
-    "crates/domains", "crates/sim_macros", 
+    "crates/domains",
+    "crates/sim_macros",
+    "crates/engine",
 ]
 resolver = "3"
 
@@ -14,10 +14,10 @@ rand = "0.9.2"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.141"
 typetag = "0.2.20"
-uuid = { version = "1.17.0", features = ["serde", "v4", "v5"] }
-thiserror = "2.0.12"
+uuid = { version = "1.18.0", features = ["serde", "v4", "v5"] }
+thiserror = "2.0.15"
 once_cell = "1.21.3"
-toml = "0.9.4"
+toml = "0.9.5"
 
 ndarray = "0.15.0"
 serde_with = "3.14.0"
@@ -32,14 +32,10 @@ tower-http = { version = "0.6", features = ["cors"] }
 
 crossbeam-channel = "0.5.15"
 
-fake = { version = "4.3.0", features = ["derive"] }
+fake = { version = "4.4.0", features = ["derive"] }
 
-proc-macro2 = "1.0"
+proc-macro2 = "1.0.101"
 quote = "1.0"
-syn = "2.0"
+syn = "2.0.106"
 
-polars = { version = "0.50.0", features = ["lazy", "csv", "serde", "parquet"] }
-linfa = { version = "0.7.1", features = ["serde"] }
-linfa-logistic = { version = "0.7.1", features = ["serde"] }
-lightgbm3 = "1.0.8"
-bincode = { version = "2.0.1", features = ["serde"] }
+
