+++ crates/domains/src/prelude/mod.rs
@@ -1,6 +1,6 @@
 pub use super::{Domain, DomainRegistration, DomainResult};
 pub use crate::banking::{BankingDomain, BankingResult, BasicBankDecisionModel};
-pub use crate::consumption::{BasicConsumerDecisionModel, CESConsumerDecisionModel, ConsumptionDomain, ConsumptionResult, ParametricMPC};
+pub use crate::consumption::{CESConsumerDecisionModel, ConsumptionDomain, ConsumptionResult};
 pub use crate::fiscal::{BasicGovernmentDecisionModel, FiscalDomain, FiscalResult};
 pub use crate::production::{BasicFirmDecisionModel, ProductionDomain, ProductionResult};
 pub use crate::settlement::{SettlementDomain, SettlementResult};
